requirements.txt
py-cord
python-dotenv
flask
gunicorn
sqlalchemy
requests